+++ libs/components/indicators/src/lib/modules/wait/types/wait-blocking-wrap-args.ts
@@ -1,5 +1,0 @@
-import { Observable } from 'rxjs';
-
-export interface SkyWaitBlockingWrapArgs<T> {
-  observable: Observable<T>;
-}